created separate ApplicationModal component
