+++ src/components/pages/PendingApplications/PendingApplication.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import axiosWithAuth from '../../../utils/axiosWithAuth';
+import ApplicationModal from './ApplicationModal';
 
 import { Table, Modal, Button } from 'antd';
 import './PendingApplication.css';
@@ -47,38 +48,12 @@ const columns = [
 const PendingApplications = () => {
   const [applications, setApplications] = useState([]);
   const [displayModal, setDisplayModal] = useState(false);
-  const [currentApplication, setCurrentApplication] = useState({});
   const [profileId, setProfileId] = useState('');
 
   const showModal = profile_id => {
     setProfileId(profile_id);
-    // getCurrentApp();
     setDisplayModal(true);
   };
-
-  const handleOk = () => {
-    setDisplayModal(false);
-    console.log(currentApplication);
-    console.log(profileId);
-  };
-
-  const handleCancel = () => {
-    setDisplayModal(false);
-  };
-
-  useEffect(() => {
-    const getCurrentApp = () => {
-      axiosWithAuth()
-        .get(`/application/profileId/${profileId}`)
-        .then(res => {
-          setCurrentApplication(res.data);
-        })
-        .catch(err => {
-          console.log(err);
-        });
-    };
-    getCurrentApp();
-  }, [profileId]);
 
   useEffect(() => {
     const getPendingApps = () => {
@@ -114,37 +89,12 @@ const PendingApplications = () => {
   return (
     <>
       <h2>Pending Applications</h2>
-
-      <Modal
-        title="Application Modal"
-        visible={displayModal}
-        onOk={handleOk}
-        onCancel={handleCancel}
-        footer={null}
-        destroyOnClose={true}
-      >
-        <h3>{`${currentApplication.first_name} ${currentApplication.last_name}`}</h3>
-        Can commit: {`${currentApplication.can_commit === true ? 'yes' : 'no'}`}
-        <br></br>
-        Submission Date: {currentApplication.created_at}
-        <br></br>
-        Current Employer: {currentApplication.current_comp}
-        <br></br>
-        Email: {currentApplication.email}
-        <br></br>
-        Availability: {currentApplication.how_commit}
-        <br></br>
-        Location: {currentApplication.location}
-        <br></br>
-        Progress Status: {currentApplication.progress_status}
-        <br></br>
-        Role: {currentApplication.role_name}
-        <br></br>
-        Tech Stack: {currentApplication.tech_stack}
-        <br></br>
-        Notes: {currentApplication.other_info}
-        <br></br>
-      </Modal>
+      <ApplicationModal
+        displayModal={displayModal}
+        setDisplayModal={setDisplayModal}
+        profileId={profileId}
+        setProfileId={setProfileId}
+      />
       <Table
         columns={columns}
         dataSource={applications}
